feat: Create TokenResponse to ensure the AppAuth library doesn't leak
- Add README.md
- Rename interface and implementation
Resolves: DCMAW-7374
