+++ build.gradle.kts
@@ -33,6 +33,7 @@ plugins {
     alias(libs.plugins.android.kotlin) apply false
     alias(libs.plugins.detekt) apply false
     alias(libs.plugins.ktlint) apply false
+    alias(libs.plugins.kotlin.serlialization) apply false
     id("org.sonarqube") version "4.3.0.3225"
     id("sonarqube-root-config")
 }
